+++ biopipen/utils/misc.R
@@ -12,3 +12,13 @@ bQuote = function(x) {
         paste0("`", x, "`")
     }
 }
+
+slugify <- function(x, non_alphanum_replace="", space_replace="_", tolower=TRUE) {
+  x <- gsub("[^[:alnum:] ]", non_alphanum_replace, x)
+  x <- trimws(x)
+  x <- gsub("[[:space:]]", space_replace, x)
+
+  if(tolower) { x <- tolower(x) }
+
+  return(x)
+}